refactor: remove additional component abstraction
